% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_model_selection.R
\name{mod_model_selection_server}
\alias{mod_model_selection_server}
\title{model selection Server Functions}
\usage{
mod_model_selection_server(id, CountryInfo, AnalysisInfo)
}
\description{
model selection Server Functions
}
